test(meshing): move 3D SphericalShell radial test to the early CI batch
SphericalShell construction is fragile under the accumulated PETSc/mesh state of
the heavy test_05*/07* CI batch (a pre-existing mesh-lifecycle issue: the
coordinate DM / cdim goes stale, giving "cannot reshape ... into shape (3)" at
build time -- unrelated to boundary slip). test_0762's SphericalShell case hit
this in CI, while test_0001 builds spheres cleanly in the early batch. Move the
3D radial-registration test to tests/test_0002_bounding_surface_3d.py so it runs
early; the Annulus tests in test_0762 cover the 2D radial logic. Verified: the
test_00[0-4]* batch (116 passed) and the test_05*/07* batch (364 passed) are
both green.

Underworld development team with AI support from Claude Code
